roots(pfr::PartialFracRational) repairs duplicated roots
When the usual `roots(pfr)` method returns duplicated roots, they might be real multiple roots. More likely, they are 2 rough estimates that happened to be in the same Newton basin of convergence and converged to be equal.

So count the number of missing roots, approximate the unsolved part of the PFR as a polynomial, and fit for it.