Change how we improve roots by Laguerre's method
Before, I was "upgrading" a (q,p>q+1) rational by creating a new partial fraction with extra terms, then computing its derivatives. But the simpler thing to do is just to drop the final (p-m) factors from the rational function and find the roots of the lower-p rational function. Those roots will be the same.